+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Changes
 - Improves the BBB integration !3050 @stcz
+- Make navigation menu sticky @AntonBallmaier @chriswalg
 - Added marker clustering to the basket map !3013 @alex.simm
 
 ## Bugfixes